refactor(PetService, PetServiceImpl, PetController, UserService): extract PetServiceImpl, centralize user lookup, and replace NoSuchElementException with EntityNotFoundException
